refactor Wallet.jsx #669
