+++ src/app/components/userProfile/wallet/WalletLine/index.js
@@ -1,0 +1,1 @@
+export { default } from './WalletLine';
