+++ src/app/components/userProfile/wallet/WalletTabs/index.js
@@ -1,0 +1,1 @@
+export { default } from './WalletTabs';
